+++ app/components/ApiSettings.vue.ts
@@ -3,7 +3,7 @@ import { Component } from 'vue-property-decorator';
 import { Inject } from '../services/core/injector';
 import GenericFormGroups from 'components/obs/inputs/GenericFormGroups.vue';
 import ObsTextInput from 'components/obs/inputs/ObsTextInput.vue';
-import { ITcpServerServiceApi, ITcpServersSettings } from '../services/tcp-server';
+import { ITcpServerServiceApi, ITcpServersSettings } from '../services/api/tcp-server';
 import { ISettingsSubCategory } from '../services/settings';
 
 @Component({
