+++ app/services/api/tcp-server/tcp-server-api.ts
@@ -1,4 +1,4 @@
-import { ISettingsSubCategory } from '../settings';
+import { ISettingsSubCategory } from 'services/settings';
 
 
 export interface ITcpServersSettings {
