+++ app/services/app/app.ts
@@ -10,8 +10,8 @@ import { SourcesService } from 'services/sources';
 import { ScenesService } from 'services/scenes';
 import { VideoService } from 'services/video';
 import { track } from 'services/usage-statistics';
-import { IpcServerService } from 'services/ipc-server';
-import { TcpServerService } from 'services/tcp-server';
+import { IpcServerService } from 'services/api/ipc-server/ipc-server';
+import { TcpServerService } from 'services/api/tcp-server';
 import { PerformanceMonitorService } from 'services/performance-monitor';
 import { SceneCollectionsService } from 'services/scene-collections';
 import { FileManagerService } from 'services/file-manager';
